// src/scheduling/appointment-types/appointment-type.ts
export interface AppointmentType {
    AppointmentTypeId: string;
    AppointmentTypeColor: string;
    DataTag: string;
    DateModified: Date;
    DefaultDuration: number;
    FormattedDuration: string;
    FontColor: string;
    Name: string;
    PerformedByProviderTypeId: number;
    UpdatesNextPreventiveAppointmentDate: boolean;
    UserModified: string;
    UsualAmount: number;
};